// frontend/src/sections/prescription-orders/detail/edit-modal/orderLineEdit.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../../api/graphql";

export type UpsertPrescriptionOrderLineVariables = {
  storeId: string;
  input: {
    id: string;
    prescriptionOrderId: string;
    itemId: string;
    quantity: number;
    note?: string | null;
  };
};

export type UpsertPrescriptionOrderLineResult = {
  upsertPrescriptionOrderLine: ({
  id: string;
});
};

export const UpsertPrescriptionOrderLine = {
  query: "mutation upsertPrescriptionOrderLine($storeId: String!, $input: UpsertPrescriptionOrderLineInput!) {\n  upsertPrescriptionOrderLine(storeId: $storeId, input: $input) {\n    ... on PrescriptionOrderLineNode {\n      id\n    }\n  }\n}",
} as TypedDocument<UpsertPrescriptionOrderLineResult, UpsertPrescriptionOrderLineVariables>;

export type DeletePrescriptionOrderLineVariables = {
  storeId: string;
  id: string;
};

export type DeletePrescriptionOrderLineResult = {
  deletePrescriptionOrderLine: ({
  id: string;
});
};

export const DeletePrescriptionOrderLine = {
  query: "mutation deletePrescriptionOrderLine($storeId: String!, $id: String!) {\n  deletePrescriptionOrderLine(storeId: $storeId, id: $id) {\n    ... on DeleteResponse {\n      id\n    }\n  }\n}",
} as TypedDocument<DeletePrescriptionOrderLineResult, DeletePrescriptionOrderLineVariables>;

export type OrderAbbreviationsVariables = Record<string, never>;

export type OrderAbbreviationsResult = {
  abbreviations: Array<{
  id: string;
  text: string;
  expansion: string;
}>;
};

export const OrderAbbreviations = {
  query: "query orderAbbreviations {\n  abbreviations {\n    id\n    text\n    expansion\n  }\n}",
} as TypedDocument<OrderAbbreviationsResult, OrderAbbreviationsVariables>;

export type OrderItemDirectionsVariables = {
  storeId: string;
  itemId: string;
};

export type OrderItemDirectionsResult = {
  items: ({
  nodes: Array<{
  id: string;
  itemDirections: Array<{
  id: string;
  directions: string;
  priority: number;
}>;
}>;
});
};

export const OrderItemDirections = {
  query: "query orderItemDirections($storeId: String!, $itemId: String!) {\n  items(storeId: $storeId, filter: {id: {equalTo: $itemId}}) {\n    ... on ItemConnector {\n      nodes {\n        id\n        itemDirections {\n          id\n          directions\n          priority\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<OrderItemDirectionsResult, OrderItemDirectionsVariables>;